+++ application_sdk/server/fastapi/models.py
@@ -1,9 +1,12 @@
 # Request/Response DTOs for workflows
 
 from enum import Enum
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List, Optional, Type
 
 from pydantic import BaseModel, Field, RootModel
+
+from application_sdk.outputs.eventstore import Event, EventFilter
+from application_sdk.workflows import WorkflowInterface
 
 
 class TestAuthRequest(RootModel[Dict[str, Any]]):
@@ -108,6 +111,19 @@ class WorkflowRequest(RootModel[Dict[str, Any]]):
         }
 
 
+class EventWorkflowRequest(BaseModel):
+    event: Event = Field(alias="data", description="Event object")
+    datacontenttype: str = Field(
+        alias="datacontenttype", description="Data content type"
+    )
+    id: str = Field(alias="id", description="Event ID")
+    source: str = Field(alias="source", description="Event source")
+    specversion: str = Field(alias="specversion", description="Event spec version")
+    time: str = Field(alias="time", description="Event time")
+    type: str = Field(alias="type", description="Event type")
+    topic: str = Field(alias="topic", description="Event topic")
+
+
 class WorkflowData(BaseModel):
     workflow_id: str = Field(..., description="Unique identifier for the workflow")
     run_id: str = Field(..., description="Unique identifier for the workflow run")
@@ -133,6 +149,17 @@ class WorkflowResponse(BaseModel):
                 },
             }
         }
+
+
+class EventWorkflowResponse(WorkflowResponse):
+    class Status(str, Enum):
+        SUCCESS = "SUCCESS"
+        RETRY = "RETRY"
+        DROP = "DROP"
+
+    # This should be a string enum of the status of the workflow, based on the Dapr docs
+    # https://docs.dapr.io/reference/api/pubsub_api/#expected-http-response
+    status: Status = Field(..., description="Status of the workflow")
 
 
 class WorkflowConfigRequest(BaseModel):
@@ -174,3 +201,23 @@ class WorkflowConfigResponse(BaseModel):
                 },
             }
         }
+
+
+class WorkflowTrigger(BaseModel):
+    workflow_class: Optional[Type[WorkflowInterface]] = None
+    model_config = {"arbitrary_types_allowed": True}
+
+
+class HttpWorkflowTrigger(WorkflowTrigger):
+    endpoint: str = "/start"
+    methods: List[str] = ["POST"]
+
+
+class EventWorkflowTrigger(WorkflowTrigger):
+    event_id: str
+    event_type: str
+    event_name: str
+    event_filters: List[EventFilter]
+
+    def should_trigger_workflow(self, event: Event) -> bool:
+        return True